+++ pubspec.yaml
@@ -12,6 +12,8 @@ dependencies:
 
   cupertino_icons: ^1.0.2
   firebase_core: ^2.24.0
+  firebase_auth: ^4.1.3
+  google_sign_in: ^6.1.6
 
 dev_dependencies:
   flutter_test:
